pyproject.toml
[project]
name = "ml-master"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.5",
    "openai>=1.98.0",
    "pandas>=2.3.1",
    "polars>=1.32.0",
    "pyarrow>=21.0.0",
    "scikit-learn>=1.7.1",
    "scipy>=1.16.1",
]
